Include new component in libraries refactor
