+++ apps/src/code-studio/components/libraries/LibraryListItem.jsx
