+++ apps/src/code-studio/components/libraries/LibraryManagerDialog.jsx
@@ -2,8 +2,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import BaseDialog from '@cdo/apps/templates/BaseDialog';
-import LibraryClientApi from '@cdo/apps/code-studio/components/Libraries/LibraryClientApi';
-import LibraryListItem from '@cdo/apps/code-studio/components/Libraries/LibraryListItem';
+import LibraryClientApi from '@cdo/apps/code-studio/components/libraries/LibraryClientApi';
+import LibraryListItem from '@cdo/apps/code-studio/components/libraries/LibraryListItem';
 import color from '@cdo/apps/util/color';
 
 const DEFAULT_MARGIN = 7;
